Fix Windows 64 bit warnings

+++ src/comm/MAVLinkSimulationWaypointPlanner.cc
@@ -890,7 +890,7 @@ void MAVLinkSimulationWaypointPlanner::mavlink_handler (const mavlink_message_t*
                 } else {
                     if (verbose) qDebug("Got MAVLINK_MSG_ID_MISSION_ITEM_REQUEST_LIST from %u but have no waypoints, staying in \n", msg->sysid);
                 }
-                protocol_current_count = waypoints->size();
+                protocol_current_count = static_cast<uint16_t>(waypoints->size());
                 send_waypoint_count(msg->sysid,msg->compid, protocol_current_count);
             } else {
                 if (verbose) qDebug("Ignored MAVLINK_MSG_ID_MISSION_ITEM_REQUEST_LIST because i'm doing something else already (state=%i).\n", current_state);
@@ -999,7 +999,7 @@ void MAVLinkSimulationWaypointPlanner::mavlink_handler (const mavlink_message_t*
                     send_waypoint_ack(protocol_current_partner_systemid, protocol_current_partner_compid, 0);
 
                     if (current_active_wp_id > waypoints_receive_buffer->size()-1) {
-                        current_active_wp_id = waypoints_receive_buffer->size() - 1;
+                        current_active_wp_id = static_cast<uint16_t>(waypoints_receive_buffer->size()) - 1;
                     }
 
                     // switch the waypoints list
